added response to voice messages
- Добавлено подключение малой модели библиотеки Vosk
- Все функции, работающие с аудиосообщениями перенесены в отдельный модуль
- Добавлено преобразование текста в аудиофайл
