+++ requirements.txt
@@ -4,3 +4,4 @@ pymongo==3.11.3
 pymodm==0.4.3
 cPython==0.0.6
 vosk==0.3.44
+pyttsx3==2.90
